separate logging directions
