--- zaifbot/bot_common/__init__.py
@@ -1,1 +1,0 @@
-
