--- zaifbot/bot_common/logging.conf
@@ -1,21 +1,0 @@
-[loggers]
-keys=root
-
-[handlers]
-keys=fileHandler
-
-[formatters]
-keys=simpleFormatter
-
-[logger_root]
-level=INFO
-handlers=fileHandler
-
-[handler_fileHandler]
-class=handlers.TimedRotatingFileHandler
-formatter=simpleFormatter
-args=(%(log_path)s,'D')
-
-[formatter_simpleFormatter]
-format=[%(asctime)s][%(levelname)s](%(filename)s:%(lineno)s) %(message)s
-datefmt=%Y/%m/%d %H:%M:%S
