+++ zaifbot/common/bot_const.py
