+++ zaifbot/common/errors.py
