+++ zaifbot/currency_pairs.py
@@ -1,8 +1,10 @@
 from threading import Thread, Event, Lock
-from .wrapper import BotPublicApi
+
 from zaifapi.impl import ZaifPublicStreamApi
-from zaifbot.bot_common.errors import ZaifBotError
-from zaifbot.bot_common.logger import logger
+from zaifbot.common.errors import ZaifBotError
+from zaifbot.common.logger import logger
+
+from .wrapper import BotPublicApi
 
 
 class CurrencyPair:
