+++ zaifbot/models/__init__.py
@@ -8,8 +8,8 @@ def _get_connection_string():
     # マルチデータベースを実現したい場合、ここを修正する
     current_dir = os.path.dirname(__file__)
     parent_dir = os.path.abspath(os.path.join(current_dir, os.path.pardir))
-    target_dir = os.path.join(parent_dir, 'bot_common')
-    return 'sqlite:///{}'.format(os.path.join(target_dir, 'zaif_bot.db'))
+    target_dir = os.path.join(parent_dir, 'common')
+    return 'sqlite:///{}'.format(os.path.join(target_dir, 'zaifbot.db'))
 
 
 Base = declarative_base()
